+++ pyproject.toml
@@ -13,6 +13,9 @@ dependencies = [
     "numpy<2.0.0",
     "matplotlib>=3.10.0",
     "torchviz>=0.0.3",
+    "ruff>=0.8.6",
+    "black>=24.10.0",
+    "mypy>=1.14.1",
 ]
 requires-python = "==3.10.*"
 readme = "README.md"
@@ -22,3 +25,5 @@ license = {text = "MIT"}
 [tool.pdm]
 distribution = false
 
+[tool.pdm.scripts]
+format = "black ."
