+++ tiny_writr/README.md
@@ -1,0 +1,16 @@
+
+   /$$     /$$                                             /$$   /$$              
+  | $$    |__/                                            |__/  | $$              
+ /$$$$$$   /$$ /$$$$$$$  /$$   /$$ /$$  /$$  /$$  /$$$$$$  /$$ /$$$$$$    /$$$$$$ 
+|_  $$_/  | $$| $$__  $$| $$  | $$| $$ | $$ | $$ /$$__  $$| $$|_  $$_/   /$$__  $$
+  | $$    | $$| $$  \ $$| $$  | $$| $$ | $$ | $$| $$  \__/| $$  | $$    | $$  \__/
+  | $$ /$$| $$| $$  | $$| $$  | $$| $$ | $$ | $$| $$      | $$  | $$ /$$| $$      
+  |  $$$$/| $$| $$  | $$|  $$$$$$$|  $$$$$/$$$$/| $$      | $$  |  $$$$/| $$      
+   \___/  |__/|__/  |__/ \____  $$ \_____/\___/ |__/      |__/   \___/  |__/      
+                         /$$  | $$                                                
+                        |  $$$$$$/                                                
+                         \______/                                                 
+
+
+# tiny_writr
+
